Fix Laravel's `runningInConsole()` on AWS Lambda
Laravel's application has a `runningInConsole()` method that returns `true` if the PHP binary used to run the process is `php-cli`.

This is a problem because it tricks Laravel into thinking that it runs the Artisan console and not the HTTP application... This can cause issues in some apps or packages.

To fix this we override the method to force its return value:
 - most of the time we don't override the value
 - when running in Lambda in HTTP we override it to return `false`