Simplify error messages for uncaught exceptions
When a handler throws an exception, it will no longer show up in the logs as `Fatal error: Uncaught <exception class>: <message> in <file>:<line>`. This gives the feeling that the developer should catch it.

Instead, developers should let exceptions bubble up. The new log message will only contain the message of the exception itself. Hopefully, that's clearer, less noisy.