Function runtime: force restarting the process for failed invocations
This is a continuation of #1013.

In the function runtime, the PHP process will now always be restarted if the invocation failed. That change only impacts apps that use `BREF_LOOP_MAX` to keep the process alive.

The reason for this is that, in most cases, a failed invocation means the PHP application threw an exception.

When using `BREF_LOOP_MAX`, the PHP process is kept alive, and it is important for the application to cleanup global state.

If an exception was thrown, it is likely the cleanup wasn't done, or wasn't done properly. It sounds much safer to restart the process.

The downside is a slight performance cost: around 10ms to restart the process. But that provides a safer execution environment.